refactor testbed
