+++ test/e2e/infra/vagrant/Vagrantfile
@@ -52,7 +52,7 @@ Vagrant.configure(VAGRANTFILE_API_VERSION) do |config|
 
   config.vm.provider "virtualbox" do |v|
     v.memory = MEMORY
-    # 2 CPUS required to initialize K8s cluster with "kubeadm init"
+    # 2 CPUS required to InitializeTestbed K8s cluster with "kubeadm init"
     v.cpus = 2
   end
