+++ test/e2e/service_externalip_test.go
@@ -473,7 +473,7 @@ func testServiceNodeFailure(t *testing.T, data *TestData) {
 			}
 			signalAgent := func(nodeName, signal string) {
 				cmd := fmt.Sprintf("pkill -%s antrea-agent", signal)
-				if testOptions.providerName != "kind" {
+				if testOptions.ProviderName != "kind" {
 					cmd = "sudo " + cmd
 				}
 				rc, stdout, stderr, err := data.RunCommandOnNode(nodeName, cmd)
